Add the same setup for android repositories than Bazel
So we can actually move the Tutorial job to be a normal job on ci.bazel.io.
This will not fix the current breakage, a parallel change is needed on ci.bazel.io.

Once this change is in, the current setup can be tested with:

```
source ci/setup_android_repositories.sh
bazel build //:all
```

Ultimately we need to do something about setting repositories for CI machines,
but I have a design doc for that :)