Update AppEngine rule to use tag 0.0.3 rather than a specific commit (#19)
The online tutorial currently directs you to use tag 0.0.2, which is
obsolete and doesn't work any more. The master branch of the tutorial
was updated to use a specific commit, but the online tutorial wasn't.

This change updates the rule to use the latest tag (0.0.3). I'll also
fix up the online tutorial page, which is in a different repo.