Add example of consuming declare_directory
This example creates a new "sum" rule in the "directory_outputs"
directory that takes a file containing a single line, a number, and adds
it to the deps inputs. Additionally there is a rule "generate_sums" that
declares a directory of files that meet this criteria. My expectation
was that it would be possible to iterate over an attribute on the File
object returned from declare_directory, but there doesn't appear to be
any way to enumerate the files that were created and pass those
individually as files into my run_shell action.

Unfortunately this doesn't seem to be the case.

I attempted to write about this in a thread on Bazel-discus
https://groups.google.com/d/msgid/bazel-discuss/CAAVj6hx7L-mbnE2qfmOuWnUrmAJxO5UJ5a8DKJqd0tgzALsmqA%40mail.gmail.com
but I figure code will do a decent job of explaining my question.

Additionally reformat with mdformat.