Add an example on how to deal with third party dependencies
There are several ways to handle external dependencies in Bazel.
However there are not many examples on how to address it or what
are the best practices.
In this example one approach is presented.
An example with third party dependencies was requested in #106.
This example has been initially developed in
https://github.com/limdor/bazel-examples, in case there is interest
other examples could be ported.