Migrate away from using the 'outputs' parameter (#216)
* expand_template: migrate away from 'outputs' parameter

The 'outputs' parameter of rule is deprecated. Hence don't use it in
examples where people are supposed to copy'n'paste from.

Link: https://docs.bazel.build/versions/main/skylark/rules.html#deprecated-predeclared-outputs
Signed-off-by: Matthias Maennich <maennich@google.com>

* actions_write: migrate away from 'outputs' parameter

The 'outputs' parameter of rule is deprecated. Hence don't use it in
examples where people are supposed to copy'n'paste from.

Link: https://docs.bazel.build/versions/main/skylark/rules.html#deprecated-predeclared-outputs
Signed-off-by: Matthias Maennich <maennich@google.com>

* computed_dependencies: migrate away from 'outputs' parameter

The 'outputs' parameter of rule is deprecated. Hence don't use it in
examples where people are supposed to copy'n'paste from.

Link: https://docs.bazel.build/versions/main/skylark/rules.html#deprecated-predeclared-outputs
Signed-off-by: Matthias Maennich <maennich@google.com>

* predeclared_outputs: migrate away from 'outputs' parameter

The 'outputs' parameter of rule is deprecated. Hence don't use it in
examples where people are supposed to copy'n'paste from.

Link: https://docs.bazel.build/versions/main/skylark/rules.html#deprecated-predeclared-outputs
Signed-off-by: Matthias Maennich <maennich@google.com>

* optional_provider: migrate away from 'outputs' parameter

The 'outputs' parameter of rule is deprecated. Hence don't use it in
examples where people are supposed to copy'n'paste from.

Link: https://docs.bazel.build/versions/main/skylark/rules.html#deprecated-predeclared-outputs
Signed-off-by: Matthias Maennich <maennich@google.com>

* depsets: migrate away from 'outputs' parameter

The 'outputs' parameter of rule is deprecated. Hence don't use it in
examples where people are supposed to copy'n'paste from.

Link: https://docs.bazel.build/versions/main/skylark/rules.html#deprecated-predeclared-outputs
Signed-off-by: Matthias Maennich <maennich@google.com>

* aspect: migrate away from 'outputs' parameter

The 'outputs' parameter of rule is deprecated. Hence don't use it in
examples where people are supposed to copy'n'paste from.

Link: https://docs.bazel.build/versions/main/skylark/rules.html#deprecated-predeclared-outputs
Signed-off-by: Matthias Maennich <maennich@google.com>

* implicit_output: migrate away from 'outputs' parameter

The 'outputs' parameter of rule is deprecated. Hence don't use it in
examples where people are supposed to copy'n'paste from.

Link: https://docs.bazel.build/versions/main/skylark/rules.html#deprecated-predeclared-outputs
Signed-off-by: Matthias Maennich <maennich@google.com>

* mandatory_provider: migrate away from 'outputs' parameter

The 'outputs' parameter of rule is deprecated. Hence don't use it in
examples where people are supposed to copy'n'paste from.

Link: https://docs.bazel.build/versions/main/skylark/rules.html#deprecated-predeclared-outputs
Signed-off-by: Matthias Maennich <maennich@google.com>