feat: extend eslint setup to full frontend 'monorepo'
Install a single eslint binary in the root of the pnpm workspace. Note, the version might not match what next.js or CRA would create in the subpackage.
Pass all eslint configs to the eslint aspect, so that it can visit any library target in the repo. ESLint should select the right configuration for each file.
In react, demonstrate using eslint as a test target rather than a build failure.