chore: move @types deps to the ts_config target
Source code doesn't have to include any indication these 'ambient' types are required.
However we can force users to be explicit in the tsconfig of which `@types` packages to include in compilation.
This follows the locality principle, since the types weren't referenced in the sources, the deps dont belong on the ts_project.