Fixed and renovated Rust example 08 grpc. (#492)
Fixed and renovaeted rust-examples/08-grpc-client-server/
Resolves:
https://github.com/bazelbuild/examples/pull/489#issuecomment-2321617323

The CI failure of the example was caused by a bizarre issue that only
occurs on linux, as documented
https://github.com/bazelbuild/rules_rust/issues/2806

I've applied the known fix. Furthermore, the cause of the issue seems to
be a change in the format of how aliases
are generated for downloaded dependencies referenced in the example. To
prevent this from ever happening again, I've vendored all dependencies
and updated the Bazel build files accordingly.

---------

Signed-off-by: Marvin Hansen <marvin.hansen@gmail.com>