Create example for usage of Symbolic Macros (#518)
This adds a new directory containing a set of interrelated symbolic
macros that builds a narrative in inline comments. It's intended as
reference material to be linked to from the official Bazel docs. It
explains core symbolic macro features, including key aspects of the
visibility system, and has a brief example of a finalizer.

Work toward bazelbuild/bazel#23857.